# src/app/api/v1/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "GapSmith Agent API",
    "version": "1.0.0",
    "description": "AI agent–accessible API for GapSmith venture intelligence. Authentication is via the x402 payment protocol — every request requires a Solana USDC payment proof in the X-PAYMENT header. Calls without payment receive HTTP 402 with payment requirements.\n\n## Two tiers\n\n- **Data API** (sub-dollar, synchronous): cached intelligence from past Scout runs. Returns results in <1s.\n- **Compute API** (few dollars, asynchronous): runs full pipelines on-demand. Returns 202 + jobId; agent polls for result.\n\n## x402 flow\n\n1. Agent makes request without `X-PAYMENT`.\n2. Server returns 402 with payment requirements (recipient ATA, amount, mint, network).\n3. Agent constructs SPL transferChecked tx, signs, submits to Solana.\n4. Agent retries the same request with `X-PAYMENT: <base64(JSON)>` header.\n5. Server verifies tx on-chain, processes request, returns 200.\n\nIdempotency: each tx_hash can only be used once. Retrying with the same `X-PAYMENT` returns the cached response (sync) or 409 (async).",
    "contact": {
      "name": "GapSmith",
      "url": "https://gapsmith.draftlabs.org"
    },
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "https://gapsmith.draftlabs.org",
      "description": "Production"
    }
  ],
  "components": {
    "securitySchemes": {
      "x402": {
        "type": "apiKey",
        "in": "header",
        "name": "X-PAYMENT",
        "description": "Base64-encoded JSON: {x402Version: 1, scheme: 'exact', network: 'solana'|'solana-devnet', payload: {txSignature: '<sig>'}}"
      }
    },
    "schemas": {
...
      "PainCluster": {
        "type": "object",
        "properties": {
          "theme": { "type": "string" },
          "description": { "type": "string" },
          "mentions": { "type": "integer" },
          "intensity": { "type": "string" }
        }
      },
...
    }
  },
  "security": [{ "x402": [] }],
  "paths": {
...
    "/api/v1/scout/pain-clusters": {
      "get": {
        "summary": "Pain cluster intelligence",
        "description": "Recurring themes from real user complaints (Reddit, HN, forums). Cached — returns instantly.",
        "tags": ["Data API"],
        "parameters": [
          { "name": "sector", "in": "query", "schema": { "type": "string" } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 20 } }
        ],
        "responses": {
          "200": {
            "description": "Pain clusters returned",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "painClusters": { "type": "array", "items": { "$ref": "#/components/schemas/PainCluster" } },
                    "count": { "type": "integer" },
                    "generatedFrom": { "type": "string", "format": "date-time", "nullable": true }
                  }
                }
              }
            }
          },
          "402": { "description": "Payment required (0.10 USDC)" }
        }
      }
    },
    "/api/v1/scout/trends": {
      "get": {
        "summary": "Trend signals",
        "description": "Emerging market shifts from recent scans. Cached — returns instantly.",
        "tags": ["Data API"],
        "parameters": [
          { "name": "sector", "in": "query", "schema": { "type": "string" } },
          { "name": "days", "in": "query", "schema": { "type": "integer", "minimum": 1, "maximum": 30, "default": 7 } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 20 } }
        ],
        "responses": {
          "200": { "description": "Trends returned" },
          "402": { "description": "Payment required (0.10 USDC)" }
        }
      }
    },
...
  }
}